refactor: prd 파이프라인 구조 개선 및 processing_status 플래그 도입
closes #3

- Repository 패턴 도입: NewsRepository Protocol + PostgresRepository/SupabaseRepository 구현체
- news_analysis_chain.py 분리: category_registry, causal_normalizer, history_builder, llm_runner
- CATEGORY_LABELS 제거: cost_categories DB 테이블 기반 동적 프롬프트 생성
- processing_status 플래그: processing→processed/failed, retry_count 관리
- src/ → prd/src/ 이동, tsconfig/vitest alias 경로 업데이트
- 불필요 파일 삭제: temp/, evals/, gemini_client.py, REFACTOR_SPEC.md, common/
- requirements.txt 운영/개발 분리 (requirements-dev.txt 추가)
