--- prd/common/__init__.py
@@ -1,2 +1,0 @@
-"""Shared helpers for the PRD package."""
-
