--- prd/evals/inputs/sample_news_politics.json
@@ -1,5 +1,0 @@
-{
-  "id": "sample-politics-001",
-  "title": "Hungary elections: what is at stake and who is challenging Viktor Orban?",
-  "content": "Hungary's upcoming election is drawing attention because opposition candidates are attempting to challenge the long-running leadership of Viktor Orban. The article focuses on domestic political stakes, campaign dynamics, and the balance of power inside Hungary. It does not describe a direct change in oil prices, exchange rates, shipping costs, import prices, or Korean household spending."
-}
