--- prd/llm/gemini_client.py
@@ -1,7 +1,0 @@
-"""Compatibility layer that exposes the graph-based analyzer."""
-
-from __future__ import annotations
-
-from prd.llm.graph.news_pipeline_graph import analyze_news
-
-__all__ = ["analyze_news"]
