+++ prd/requirements-dev.txt
@@ -1,0 +1,3 @@
+-r requirements.txt
+pytest>=8.0.0
+pytest-asyncio>=0.23.0
