+++ prd/requirements.txt
@@ -4,7 +4,5 @@ langgraph>=0.2.0
 pydantic>=2.8.0
 psycopg[binary]==3.2.7
 python-dotenv==1.0.1
-pytest>=8.0.0
-pytest-asyncio>=0.23.0
 supabase>=2.3.0,<3
