+++ prd/src/lib/lagCalculator.ts
