--- prd/temp/keyword_report.txt
@@ -1,105 +1,0 @@
-전체 뉴스 건수: 57
-
-============================================================
-keyword (상위 50개)
-============================================================
-  war                                57건  (100.0%)
-  price                              44건  ( 77.2%)
-  oil                                39건  ( 68.4%)
-  conflict                           35건  ( 61.4%)
-  energy                             32건  ( 56.1%)
-  fuel                               31건  ( 54.4%)
-  middle east                        31건  ( 54.4%)
-  cost                               29건  ( 50.9%)
-  military                           27건  ( 47.4%)
-  gas                                25건  ( 43.9%)
-  attack                             23건  ( 40.4%)
-  inflation                          18건  ( 31.6%)
-  russia                             15건  ( 26.3%)
-  shipping                           14건  ( 24.6%)
-  ukraine                            14건  ( 24.6%)
-  invasion                           11건  ( 19.3%)
-  food                               11건  ( 19.3%)
-  gaza                                4건  (  7.0%)
-  geopolitical                        4건  (  7.0%)
-  commodity                           4건  (  7.0%)
-  wheat                               1건  (  1.8%)
-
-============================================================
-increased_items (상위 30개)
-============================================================
-  price                              23건  ( 40.4%)
-  oil                                22건  ( 38.6%)
-  cost                               13건  ( 22.8%)
-  energy                             12건  ( 21.1%)
-  inflation                          12건  ( 21.1%)
-  fuel                               11건  ( 19.3%)
-  gas                                11건  ( 19.3%)
-  food                                7건  ( 12.3%)
-  shipping                            3건  (  5.3%)
-  commodity                           3건  (  5.3%)
-
-============================================================
-decreased_items (상위 30개)
-============================================================
-  price                               6건  ( 10.5%)
-  energy                              4건  (  7.0%)
-  oil                                 3건  (  5.3%)
-  gas                                 2건  (  3.5%)
-  shipping                            2건  (  3.5%)
-  food                                1건  (  1.8%)
-  inflation                           1건  (  1.8%)
-  cost                                1건  (  1.8%)
-  fuel                                1건  (  1.8%)
-
-============================================================
-전체 keyword 목록 (알파벳순)
-============================================================
-  attack                             23건
-  commodity                           4건
-  conflict                           35건
-  cost                               29건
-  energy                             32건
-  food                               11건
-  fuel                               31건
-  gas                                25건
-  gaza                                4건
-  geopolitical                        4건
-  inflation                          18건
-  invasion                           11건
-  middle east                        31건
-  military                           27건
-  oil                                39건
-  price                              44건
-  russia                             15건
-  shipping                           14건
-  ukraine                            14건
-  war                                57건
-  wheat                               1건
-
-============================================================
-전체 increased_items 목록 (알파벳순)
-============================================================
-  commodity                           3건
-  cost                               13건
-  energy                             12건
-  food                                7건
-  fuel                               11건
-  gas                                11건
-  inflation                          12건
-  oil                                22건
-  price                              23건
-  shipping                            3건
-
-============================================================
-전체 decreased_items 목록 (알파벳순)
-============================================================
-  cost                                1건
-  energy                              4건
-  food                                1건
-  fuel                                1건
-  gas                                 2건
-  inflation                           1건
-  oil                                 3건
-  price                               6건
-  shipping                            2건
